add pausable contracts, revert reasons, and finalized contract registry
## EIP820 official repository has been added as a submodule to our contracts

This will allow us to contribute directly to the official [ERC-820 standard](https://github.com/jbaylina/eip820/) when appropriate. It will now be much easier to also grab the latest updates.

If you want to contribute/modify the 820 contracts. See the README in `contracts/EIP820/README.md`

* initial erc820 refactor to use submodule

* add submodule to contract repo

* update readme with submodule info

* add submodule init to circle

* remove old erc820 imports

## Reason string support for reverts 

* add reason to all the reverts

When truffle reverts, we will now be able to see why! Unfortunately not until we migrate to truffle v5, however-- there is a ticket open for this. 


## Pausability in contracts 

* add ability to pause contracts (#69)

* add pausable contract funcs and onlyOwner (#71)


## Contract Registry

* finalize contract registry (#72)

* add more contract base functionality

This adds a significant number of additional functions to look up versions. Hopefully this means the contract and root registry will not need to be nuked for a long time. ENS will be updated to resolve to the new root.

`getContractNameAndHashAtProxy` to return the contract name and it's hash being used by a particular proxy address

`getVersionCountForContract` Gets the number of Versions for a particular ContractName at a particular proxy address

`getContractInfoForVersion` Gets a contract version info for a particular contract name and proxy using the version name, which returns the index at which the given contract exists, the name of the version, the logic implementation, and the address of the proxy used by this versions parent

`getVersionHistoryForContractName` returns the entire version history for a contract 

`setVersionAsAdmin` Allows the MultiAdmin to set a version manually. 

`setVersion` can now only be called by the proxy contract

`registeredContractCount.call()` will now return the total number of registered contracts


* finalize registry and root

* add tests for all registry functionality

* change registry interface functions to external



## Misc

* ignore truffles tmp generated .node-xmlhttp files

* various test clean up

* call migration interface toggles using admin  (#70)

* add token init event

* Use more recent `Ownable` contract, remove `UnstructuredOwnable` nomenclature

* add setter functions for introspection registry and eip820 registry

* bump versions:

 `ContractRegistryV0_2_0` -> `ContractRegistryV0_3_0`

 `FifoCrcMarketV0_1_2` -> `FifoCrcMarketV0_2_0`

 `ParticipantRegistryV0_1_2` -> `ParticipantRegistryV0_2_0`

 `ParticipantV0_2_0` -> `ParticipantV0_3_0`

 `SupplierV0_2_0` -> `SupplierV0_3_0`

 `VerifierV0_2_0` -> `VerifierV0_3_0`

 `NoriV0_1_2` -> `NoriV0_2_0`

 `CRCV0_2_1` -> `CRCV0_3_0`

* add latest docs


#### full commit list

* add reason to all the reverts

* bump versions, need to fix tests accordingly

* fix test

* fix fifo version name

* fix tests

* initial erc820 refactor to use submodule

* add submodule to contract repo vs mono

* make tests work

* update submodule

* update readme with submodule info

* add submodule init to circle

* add missing dep

* fix submodule

* ignore truffles tmp generated .node-xmlhttp files

* remove test token extension

* remove old erc820 imports

* clean up

* remove old 820 files

* [In progress] add ability to pause contracts (#69)

* add pausable nori example

* call toggles using admin (#70)

* call toggles using admin

* add pausable contract funcs and onlyOwner (#71)

* finalize contract registry (#72)

* checkpoint

* finalize registry

* add token init event back

* add more contract base functionality

* finalize registry and root

* give admin account more eth for tests

* add more registry base tests

* add final tests and functions for registry

* rename version registry tests to contract registry base tests

* change registry interface functions to external

* add latest docs, bump contract versions